move heartbeat listeners to leader state
